refactor: extract data fetching from podcast category section
